pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Str1kez <vazinov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
environs = "^9.5.0"
dj-database-url = "^1.0.0"
psycopg2 = "^2.9.3"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.1"}
Pillow = "^9.2.0"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.24.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
